+++ ignite/pkg/cosmosclient/consensus.go
@@ -5,8 +5,8 @@ import (
 	"encoding/base64"
 	"time"
 
+	tmproto "github.com/cometbft/cometbft/api/cometbft/types/v1"
 	"github.com/cometbft/cometbft/libs/bytes"
-	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	tmtypes "github.com/cometbft/cometbft/types"
 )