+++ ignite/templates/typed/simapp.go
@@ -48,9 +48,9 @@ func ModuleSimulationMsgModify(
 			templateOpMsg := `simulation.NewWeightedProposalMsg(
 	opWeightMsg%[2]v%[3]v,
 	defaultWeightMsg%[2]v%[3]v,
-	func(r *rand.Rand, ctx sdk.Context, accs []simtypes.Account) sdk.Msg {
+	func(r *rand.Rand, accs []simtypes.Account, cdc address.Codec) (sdk.Msg, error) {
 		%[4]vsimulation.SimulateMsg%[2]v%[3]v(am.accountKeeper, am.bankKeeper, am.keeper, simState.TxConfig)
-		return nil
+		return nil, nil
 	},
 ),
 %[1]v`