+++ ignite/pkg/cosmosutil/gentx.go
@@ -6,16 +6,28 @@ import (
 	"os"
 
 	sdkmath "cosmossdk.io/math"
+<<<<<<< HEAD
 	"github.com/cometbft/cometbft/crypto/ed25519"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/ignite/cli/v28/ignite/pkg/errors"
+=======
+
+	"github.com/cometbft/cometbft/crypto/ed25519"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/ignite/cli/v29/ignite/pkg/errors"
+>>>>>>> bf2539d6 (fix: fix gentx parser into the cosmosutil package (#4402))
 )
 
 type (
 	// GentxInfo represents the basic info about gentx file.
 	GentxInfo struct {
+		// Deprecated: Use of Delegator Address in MsgCreateValidator is deprecated.
+		// The validator address bytes and delegator address bytes refer to the same account while creating validator (defer
+		// only in bech32 notation).
 		DelegatorAddress string
+		ValidatorAddress string
 		PubKey           ed25519.PubKey
 		SelfDelegation   sdk.Coin
 		Memo             string
@@ -73,6 +85,7 @@ func ParseGentx(gentxBz []byte) (info GentxInfo, err error) {
 
 	info.Memo = gentx.Body.Memo
 	info.DelegatorAddress = gentx.Body.Messages[0].DelegatorAddress
+	info.ValidatorAddress = gentx.Body.Messages[0].ValidatorAddress
 
 	pb := gentx.Body.Messages[0].PubKey.Key
 	info.PubKey, err = base64.StdEncoding.DecodeString(pb)