+++ changelog.md
@@ -43,6 +43,7 @@
 - [#4317](https://github.com/ignite/cli/pull/4317) Remove xchisel dependency
 - [#4361](https://github.com/ignite/cli/pull/4361) Remove unused `KeyPrefix` method
 - [#4376](https://github.com/ignite/cli/pull/4376) Set different chain-id for in place testnet
+- [#4384](https://github.com/ignite/cli/pull/4384) Compare genesis params into chain genesis tests
 
 ### Fixes