+++ changelog.md
@@ -9,6 +9,14 @@
 ### Changes
 
 - [#4376](https://github.com/ignite/cli/pull/4376) Set different chain-id for in place testnet
+- [#4384](https://github.com/ignite/cli/pull/4384) Compare genesis params into chain genesis tests
+
+### Fixes
+
+- [#4000](https://github.com/ignite/cli/pull/4000) Run all dry runners before the wet run in the `xgenny` pkg
+- [#4091](https://github.com/ignite/cli/pull/4091) Fix race conditions in the plugin logic
+- [#4128](https://github.com/ignite/cli/pull/4128) Check for duplicate proto fields in config
+- [#4402](https://github.com/ignite/cli/pull/4402) Fix gentx parser into the cosmosutil package
 
 ## [`v28.5.3`](https://github.com/ignite/cli/releases/tag/v28.5.3)