+++ changelog.md
@@ -21,7 +21,7 @@
 - [#4377](https://github.com/ignite/cli/pull/4377) Add multi node (validator) testnet.
 - [#4326](https://github.com/ignite/cli/pull/4326) Add `buf.build` version to `ignite version` command
 - [#4362](https://github.com/ignite/cli/pull/4362) Scaffold `Makefile`
-- [#4289](https://github.com/ignite/cli/pull/4289) Cosmos SDK v0.52 support
+- [#4289](https://github.com/ignite/cli/pull/4289), [#4423](https://github.com/ignite/cli/pull/4423) Cosmos SDK v0.52 support
 
 ### Changes